A11y updates (#9)
* Adjust outline offsets.

* Use a different ARIA label for thumbnail diagram SVGs.

* Make pendulum arm keyboard focusable and clickable.

* Improve keyboard focusability by preventing outline on click, and make fig. 18 options keyboard focusable.

* Add fill animation to pendulum click overlay to better indicate clickability.

* Add aria-live, aria-selected and aria-current to the semantics between the toolbar and the figure.

* Hide the animated caption from screen readers and add an accessible, visually hidden caption.

The animated caption, even with role="text", was being read out letter
by letter.

Resources for the current approach:

- https://css-irl.info/how-to-accessibly-split-text/
- https://kittygiraudel.com/2021/02/17/hiding-content-responsibly/?ref=uxcollective

* Bump up max diagram size.

Fig. 18 was quite hard to read on a large desktop screen.

* Use aria-current instead of aria-selected for figure toolbar.

aria-selected implies role=tab and role=tablist, which, in turn,
required each tab item to aria-control a specific tab. We don't render
each figure immediately and switch visibility via tabbing, so it's not
possible to add the correct aria roles. In any case, since these are
anchor tags that change the URL, aria-selected may be more appropriate
anyway.

* Try out native scrollIntoView for fig. 18 boxes.

Reasons not to go with this:
- Smooth option not supported on Safari
- Not in sync with the arrow drawing, so does it actually help?
- Interferes with manual scrolling?

* Revert "Try out native scrollIntoView for fig. 18 boxes."

This reverts commit e1a9c9345911731f1fd832bd6799da2a530037ca.

Too much effort to get to work nicely.

* Update CHANGELOG.